+++ tests/NCronJob.Tests/IntegrationTests.cs
@@ -147,7 +147,7 @@ public sealed class IntegrationTests : JobIntegrationBase
     [Theory]
     [MemberData(nameof(InstantJobRunners))]
     public async Task InstantJobCanStartADisabledJob(
-        Func<IInstantJobRegistry, TimeProvider, string, CancellationToken, Guid> instantJobRunner)
+        Func<IInstantJobRegistry, TimeProvider, object?, CancellationToken, Guid> instantJobRunner)
     {
         ServiceCollection.AddNCronJob(
             n => n.AddJob<DummyJob>((jo) => jo.WithCronExpression(Cron.AtEveryMinute)));
@@ -176,7 +176,7 @@ public sealed class IntegrationTests : JobIntegrationBase
     [Theory]
     [MemberData(nameof(InstantJobRunners))]
     public async Task ShouldThrowRuntimeExceptionWhenTriggeringThroughTheInstantJobRegistryAnAmbiguousTypeReference(
-        Func<IInstantJobRegistry, TimeProvider, string, CancellationToken, Guid> instantJobRunner)
+        Func<IInstantJobRegistry, TimeProvider, object?, CancellationToken, Guid> instantJobRunner)
     {
         ServiceCollection.AddNCronJob(n =>
         {
@@ -193,16 +193,147 @@ public sealed class IntegrationTests : JobIntegrationBase
         Action act = () => instantJobRunner(instantJobRegistry, FakeTimer, "Hello from InstantJob", CancellationToken);
 
         act.ShouldThrow<InvalidOperationException>()
-            .Message.ShouldContain($"Ambiguous job reference for type 'DummyJob' detected.");
+            .Message.ShouldContain("Ambiguous job reference for type 'DummyJob' detected.");
     }
 
-    public static TheoryData<Func<IInstantJobRegistry, TimeProvider, string, CancellationToken, Guid>> InstantJobRunners()
+    public static TheoryData<Func<IInstantJobRegistry, TimeProvider, object?, CancellationToken, Guid>> InstantJobRunners()
     {
-        var t = new TheoryData<Func<IInstantJobRegistry, TimeProvider, string, CancellationToken, Guid>>();
+        var t = new TheoryData<Func<IInstantJobRegistry, TimeProvider, object?, CancellationToken, Guid>>();
         t.Add((i, f, p, t) => i.RunInstantJob<DummyJob>(p, t));
         t.Add((i, f, p, t) => i.RunScheduledJob<DummyJob>(f.GetUtcNow(), p, t));
         t.Add((i, f, p, t) => i.ForceRunInstantJob<DummyJob>(p, t));
         t.Add((i, f, p, t) => i.ForceRunScheduledJob<DummyJob>(TimeSpan.Zero, p, t));
+        return t;
+    }
+
+    [Theory]
+    [MemberData(nameof(InstantNamedJobRunners))]
+    public async Task CanDisambiguateSimarlyTypedJobsThroughNames(
+        Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid> instantJobRunner)
+    {
+        ServiceCollection.AddNCronJob(n =>
+        {
+            n.AddJob<DummyJob>(s => s.WithName("good").WithParameter("good_param"))
+                .ExecuteWhen(success: s => s.RunJob<AnotherDummyJob>());
+            n.AddJob<DummyJob>(s => s.WithCronExpression(Cron.Never).WithParameter("bad_param"))
+                .ExecuteWhen(success: s => s.RunJob<ExceptionJob>());
+        });
+
+        await StartNCronJob(startMonitoringEvents: true);
+
+        var instantJobRegistry = ServiceProvider.GetRequiredService<IInstantJobRegistry>();
+
+        var orchestrationId = instantJobRunner(instantJobRegistry, FakeTimer, "good", null, CancellationToken);
+
+        await WaitForOrchestrationCompletion(orchestrationId, stopMonitoringEvents: true);
+
+        Storage.Entries[0].ShouldBe("DummyJob - Parameter: good_param");
+        Storage.Entries[1].ShouldBe("AnotherDummyJob - Parameter: ");
+        Storage.Entries.Count.ShouldBe(2);
+    }
+
+    [Theory]
+    [MemberData(nameof(InstantNamedJobRunners))]
+    public async Task CanDisambiguateSimarlyTypedJobsThroughNamesWhileOverridingParameters(
+        Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid> instantJobRunner)
+    {
+        ServiceCollection.AddNCronJob(n =>
+        {
+            n.AddJob<DummyJob>(s => s.WithName("good").WithParameter("good_param"))
+                .ExecuteWhen(success: s => s.RunJob<AnotherDummyJob>());
+            n.AddJob<DummyJob>(s => s.WithCronExpression(Cron.Never).WithParameter("bad_param"))
+                .ExecuteWhen(success: s => s.RunJob<ExceptionJob>());
+        });
+
+        await StartNCronJob(startMonitoringEvents: true);
+
+        var instantJobRegistry = ServiceProvider.GetRequiredService<IInstantJobRegistry>();
+
+        var orchestrationId = instantJobRunner(instantJobRegistry, FakeTimer, "good", "overriden_param", CancellationToken);
+
+        await WaitForOrchestrationCompletion(orchestrationId, stopMonitoringEvents: true);
+
+        Storage.Entries[0].ShouldBe("DummyJob - Parameter: overriden_param");
+        Storage.Entries[1].ShouldBe("AnotherDummyJob - Parameter: ");
+        Storage.Entries.Count.ShouldBe(2);
+    }
+
+    [Theory]
+    [MemberData(nameof(InstantNamedJobRunners))]
+    public async Task CanTrigerDynamicJobsThroughNames(
+        Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid> instantJobRunner)
+    {
+        ServiceCollection.AddNCronJob(n =>
+        {
+            n.AddJob(
+                dynamicJob,
+                Cron.Never,
+                jobName: "good");
+        });
+
+        await StartNCronJob(startMonitoringEvents: true);
+
+        var instantJobRegistry = ServiceProvider.GetRequiredService<IInstantJobRegistry>();
+
+        var orchestrationId = instantJobRunner(instantJobRegistry, FakeTimer, "good", null, CancellationToken);
+
+        await WaitForOrchestrationCompletion(orchestrationId, stopMonitoringEvents: true);
+
+        Storage.Entries[0].ShouldBe("Done - Parameter : ");
+        Storage.Entries.Count.ShouldBe(1);
+    }
+
+    [Theory]
+    [MemberData(nameof(InstantNamedJobRunners))]
+    public async Task CanTrigerDynamicJobsThroughNamesWhilePassingParameters(
+        Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid> instantJobRunner)
+    {
+        ServiceCollection.AddNCronJob(n =>
+        {
+            n.AddJob(
+                dynamicJob,
+                Cron.Never,
+                jobName: "good");
+        });
+
+        await StartNCronJob(startMonitoringEvents: true);
+
+        var instantJobRegistry = ServiceProvider.GetRequiredService<IInstantJobRegistry>();
+
+        var orchestrationId = instantJobRunner(instantJobRegistry, FakeTimer, "good", "param", CancellationToken);
+
+        await WaitForOrchestrationCompletion(orchestrationId, stopMonitoringEvents: true);
+
+        Storage.Entries[0].ShouldBe("Done - Parameter : param");
+        Storage.Entries.Count.ShouldBe(1);
+    }
+
+    [Theory]
+    [MemberData(nameof(InstantNamedJobRunners))]
+    public async Task ShouldThrowRuntimeExceptionWhenTriggeringThroughTheInstantJobRegistryAnUnregisteredNamedJob(
+        Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid> instantJobRunner)
+    {
+        ServiceCollection.AddNCronJob(n =>
+        {
+        });
+
+        await StartNCronJob(startMonitoringEvents: false);
+
+        var instantJobRegistry = ServiceProvider.GetRequiredService<IInstantJobRegistry>();
+
+        Action act = () => instantJobRunner(instantJobRegistry, FakeTimer, "good", "param", CancellationToken);
+
+        act.ShouldThrow<InvalidOperationException>()
+            .Message.ShouldContain("Job with name 'good' not found.");
+    }
+
+    public static TheoryData<Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid>> InstantNamedJobRunners()
+    {
+        var t = new TheoryData<Func<IInstantJobRegistry, TimeProvider, string, object?, CancellationToken, Guid>>();
+        t.Add((i, f, n, p, t) => i.RunInstantJob(n, p, t));
+        t.Add((i, f, n, p, t) => i.RunScheduledJob(n, f.GetUtcNow(), p, t));
+        t.Add((i, f, n, p, t) => i.ForceRunInstantJob(n, p, t));
+        t.Add((i, f, n, p, t) => i.ForceRunScheduledJob(n, TimeSpan.Zero, p, t));
         return t;
     }
 
@@ -684,6 +815,9 @@ public sealed class IntegrationTests : JobIntegrationBase
         filteredEvents.ShouldBeScheduledThenCompleted();
         return true;
     }
+
+    private readonly Delegate dynamicJob = (IJobExecutionContext context, Storage storage, CancellationToken token) 
+        => { storage.Add($"Done - Parameter : {context.Parameter}"); };
 
     private static class JobMethods
     {